Refactor.  This is a rather large change.  I separated out the chart drawing from the data structure manipulation.  This definitely makes our code more modular.
The big win is that we now create a new chart every time we switch schemes,
meaning that the stuff in the legend is actually reasonable now.

Definitely needs some thorough testing, since the code has changed so much.
I think it looks OK, but another pair of eyes (or two, or three) certainly
can't hurt.